Part 1

+++ dev/Presentation/RoleScreen.java
@@ -1,6 +1,5 @@
 package Presentation;
 
-import Domain.Role;
 import Service.RoleController;
 
 import java.util.Scanner;
@@ -18,22 +17,42 @@ public class RoleScreen {
         int choose = -1;
         String roleName;
         Scanner scanner = new Scanner(System.in);
-        System.out.println("1)Create Role\n2)Delete Role\n3)Go Back\n");
         while (choose != 3) {
+            System.out.println("1)Create Role\n2)Delete Role\n3)Go Back");
             choose = scanner.nextInt();
             switch (choose) {
                 case 1:
+                    //
                     System.out.println("Enter Role Name: ");
                     roleName = scanner.next();
-                    // add role to DB.
-                    rc.addRole(roleName);
+                    if (rc.getRole(roleName) == null)
+                        // add role to DB.
+                        rc.addRole(roleName);
+                    else System.out.println("Role already exists");
                     break;
                 case 2:
-                    // get the index.
+                    // Delete Role
+                    System.out.println("Enter Role Name");
                     roleName = scanner.next();
+                    if (rc.getRole(roleName) != null)
+                        rc.deleteRole(roleName);
+                    else System.out.println("Role does not exist");
                     // delete the specific role from db.
-                    rc.deleteRole(roleName);
                     break;
+ /*    MODIFYING CANT DO's            case 3:
+                    // Modify Role Cant DO's.
+                    System.out.println("Enter Existing Role Name");
+                    roleName = scanner.next();
+                    if (rc.getRole(roleName) != null) {
+                        System.out.println("Enter a role that, " + roleName + "cant do:");
+                        String cantDo = scanner.next();
+                        if (rc.getRole(cantDo) != null)
+                            rc.getRole(roleName).addCantDo(rc.getRole(cantDo));
+                        else System.out.println("Role does not exist");
+                    } else System.out.println("Role does not exist");
+
+                    // delete the specific role from db.
+                    break;*/
                 case 3:
                     break;
                 default:
@@ -42,6 +61,8 @@ public class RoleScreen {
             }
         }
     }
+
+
 }
